+++ src/components/Bagel/Bagel.styled.js
@@ -1,5 +1,14 @@
 import { styled } from 'styled-components';
 
 export const BagelStyled = styled.div`
-  background-color: transparent;
+  width: 280px;
+  height: 280px;
+  @media only screen and (min-width: 768px) {
+    width: 336px;
+    height: 336px;
+  }
+  @media only screen and (min-width: 1280px) {
+    width: 288px;
+    height: 288px;
+  }
 `;
