+++ src/components/TransactionsList/TransactionsList.jsx
@@ -1,6 +1,4 @@
-import { SpriteSVG } from 'pictures/SpriteSVG';
 import React, { useEffect } from 'react';
-import moment from 'moment';
 import { useDispatch, useSelector } from 'react-redux';
 import {
   deleteTransaction,
@@ -13,44 +11,12 @@ import {
 import { ModalEditTransaction } from 'components/ModalEditTransaction/ModalEditTransaction';
 import Modal from 'components/Modal/Modal';
 import { isEditTransaction } from 'redux/Global/selectors';
-import {
-  DeleteTabBtn,
-  EditIconStyled,
-  IconBtnWrapperStyled,
-  TableStyled,
-  TdActionStyled,
-  TdCatagoryStyled,
-  TdDateStyled,
-  TdCommentStyled,
-  TdSumStyled,
-  TdTypeStyled,
-  TrInfoStyled,
-  ThDateStyled,
-  ThSumStyled,
-  TableHeader,
-  TableBody,
-  TableScrollBody,
-  ThTypeStyled,
-  ThCategoryStyled,
-  ThCommentStyled,
-  ThActionStyled,
-  TrStyled,
-} from './TransactionsList.styled';
 import { selectAllCategories } from 'redux/TransactionCategories/selectors';
-import { getCategoriName } from 'helpers/helpers';
 import { feachCategories } from 'redux/TransactionCategories/operations';
-import { setIsModalEditTransaction } from 'redux/Global/globalSlice';
 import { useMediaQuery } from 'react-responsive';
 import { IS_DESKTOP, IS_MOBILE, IS_TABLET } from 'styles/const ';
-import {
-  CommentMobTrans,
-  DeletBtnMobTrans,
-  EditBtnMobTrans,
-  EditIconMobTrans,
-  StyledLiTransList,
-  StyledNameTranList,
-  StyledUlTransList,
-} from './TransactionsList.styled';
+import MobileList from './MobileList';
+import TabletAndDesctopList from './TabletAndDesctopList';
 
 const TransactionsList = () => {
   const isMobile = useMediaQuery(IS_MOBILE);
@@ -77,115 +43,20 @@ const TransactionsList = () => {
 
   return (
     <>
-      {isMobile &&
-        dataList?.map(
-          ({ id, transactionDate, type, categoryId, comment, amount }) => (
-            <StyledUlTransList key={id}>
-              <StyledLiTransList>
-                <StyledNameTranList>Date</StyledNameTranList>
-                <div>{moment(transactionDate).format('DD.MM.YY')}</div>
-              </StyledLiTransList>
-              <StyledLiTransList>
-                <StyledNameTranList>Type</StyledNameTranList>
-                <div>{type === 'INCOME' ? '+' : '-'}</div>
-              </StyledLiTransList>
-              <StyledLiTransList>
-                <StyledNameTranList>Category</StyledNameTranList>
-                <div>
-                  {dataList.length !== 0
-                    ? getCategoriName(allCategories, categoryId)
-                    : 'notNull'}
-                </div>
-              </StyledLiTransList>
-              <StyledLiTransList>
-                <StyledNameTranList>Comment</StyledNameTranList>
-                <CommentMobTrans>{comment}</CommentMobTrans>
-              </StyledLiTransList>
-              <StyledLiTransList>
-                <StyledNameTranList>Sum</StyledNameTranList>
-                <div>{Math.abs(amount)}</div>
-              </StyledLiTransList>
-              <StyledLiTransList>
-                <DeletBtnMobTrans onClick={() => handleDeleteClick(id)}>
-                  Delete
-                </DeletBtnMobTrans>
-                <EditBtnMobTrans
-                  onClick={() => {
-                    dispatch(setIsModalEditTransaction({ id, flag: true }));
-                  }}
-                >
-                  <EditIconMobTrans>
-                    <SpriteSVG name={'edit'} />
-                  </EditIconMobTrans>
-                  Edit
-                </EditBtnMobTrans>
-              </StyledLiTransList>
-            </StyledUlTransList>
-          )
-        )}
-      {(isTablet || isDesktop) && (
-        <>
-          <TableStyled>
-            <TableHeader>
-              <TrStyled>
-                <ThDateStyled>Date</ThDateStyled>
-                <ThTypeStyled>Type</ThTypeStyled>
-                <ThCategoryStyled>Category</ThCategoryStyled>
-                <ThCommentStyled>Comment</ThCommentStyled>
-                <ThSumStyled>Sum</ThSumStyled>
-                <ThActionStyled>&nbsp;</ThActionStyled>
-              </TrStyled>
-            </TableHeader>
-          </TableStyled>
-          <TableScrollBody>
-            <TableBody>
-              {dataList?.map(
-                ({
-                  id,
-                  transactionDate,
-                  type,
-                  categoryId,
-                  comment,
-                  amount,
-                }) => (
-                  <TrInfoStyled key={id}>
-                    <TdDateStyled>
-                      {moment(transactionDate).format('DD.MM.YY')}
-                    </TdDateStyled>
-                    <TdTypeStyled>{type === 'INCOME' ? '+' : '-'}</TdTypeStyled>
-                    <TdCatagoryStyled>
-                      {dataList.length !== 0
-                        ? getCategoriName(allCategories, categoryId)
-                        : 'notNull'}
-                    </TdCatagoryStyled>
-                    <TdCommentStyled>{comment}</TdCommentStyled>
-                    <TdSumStyled>{Math.abs(amount)}</TdSumStyled>
-                    <TdActionStyled>
-                      <IconBtnWrapperStyled>
-                        <div
-                          onClick={() => {
-                            dispatch(
-                              setIsModalEditTransaction({ id, flag: true })
-                            );
-                          }}
-                        >
-                          <EditIconStyled>
-                            <SpriteSVG name={`edit`} />
-                          </EditIconStyled>
-                        </div>
-                        <DeleteTabBtn onClick={() => handleDeleteClick(id)}>
-                          Delete
-                        </DeleteTabBtn>
-                      </IconBtnWrapperStyled>
-                    </TdActionStyled>
-                  </TrInfoStyled>
-                )
-              )}
-            </TableBody>
-          </TableScrollBody>
-        </>
+      {isMobile && (
+        <MobileList
+          dataList={dataList}
+          allCategories={allCategories}
+          handleDeleteClick={handleDeleteClick}
+        />
       )}
-
+      {(isTablet || isDesktop) && (
+        <TabletAndDesctopList
+          dataList={dataList}
+          allCategories={allCategories}
+          handleDeleteClick={handleDeleteClick}
+        />
+      )}
       {isEditTrans && (
         <Modal>
           <ModalEditTransaction />
