+++ src/pages/SummaryPage/SummaryPage.styled.js
@@ -6,11 +6,14 @@ export const SummaryPageContStyled = styled.div`
   margin: 0 auto;
   display: flex;
   flex-direction: column;
+  /* justify-content: center; */
+  padding: 28px 0 0 0;
   gap: 32px;
   width: 320px;
   @media screen and (min-width: 768px) {
-    width: 768px;
+    width: 704px;
     flex-direction: row;
+    padding: 20px 0 0 0;
   }
   @media screen and (min-width: 1280px) {
     width: 800px;
